Added auto cleanup functionality for Oh-My-Zsh
This commit enhances the Oh-My-Zsh configuration by introducing an automatic cleanup feature. Now, if any plugins are removed from the repos array, their associated plugin folders in the $ZSH_CUSTOM/plugins directory will be deleted. This ensures that the configuration remains updated and in line with the defined set of plugins.